+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Adding `c::stripLeadingIndents()`.
 - Refactoring `Classes\Tokenizer{}`.
+- Making it possible to tokenize specific shortcode tag names in `Classes\Core\Tokenizer{}`.
 
 ## v160719.52827